simplify deployment of whole project on localhost
Adapt deployment on localhost so that no manual adaptions have to be
done (no more need to change URL for backend API inside Angular).
For this purpose, rewrite Docker Compose config to include an NGINX
which acts as a reverse proxy. The NGINX serves both applications
(frontend and backend) on localhost on the same port (80).
Update README to reflect these changes.